refactor: use case and tests

+++ src/Applications/use_case/_test/DeleteThreadCommentUseCase.test.js
@@ -13,15 +13,11 @@ describe('DeleteThreadCommentUseCase', () => {
     // mock
     const mockThreadRepository = new ThreadRepository();
     const date = '2021-08-08T07:22:13.017Z';
-    mockThreadRepository.getCommentById = jest.fn(() => Promise.resolve({
-      owner: 'user-123',
-      threadId: 'thread-123',
-      deletedDate: null,
-    }));
-    mockThreadRepository.getThreadById = jest.fn(() => Promise.resolve({
-      id: 'thread-123',
-    }));
-    mockThreadRepository.deleteCommentById = jest.fn(() => Promise.resolve(true));
+
+    mockThreadRepository.verifyThreadAvailability = jest.fn(() => true);
+    mockThreadRepository.verifyCommentAvailability = jest.fn(() => true);
+    mockThreadRepository.verifyCommentOwnership = jest.fn(() => true);
+    mockThreadRepository.deleteCommentById = jest.fn(() => true);
 
     const dateNowSpy = jest.spyOn(Date.prototype, 'toISOString');
     dateNowSpy.mockImplementationOnce(() => date);
@@ -36,8 +32,11 @@ describe('DeleteThreadCommentUseCase', () => {
     dateNowSpy.mockRestore();
 
     // Assert
-    expect(mockThreadRepository.getCommentById).toBeCalledWith(useCasePayload.commentId);
-    expect(mockThreadRepository.getThreadById).toBeCalled();
+    expect(mockThreadRepository.verifyThreadAvailability).toBeCalledWith(useCasePayload.threadId);
+    expect(mockThreadRepository.verifyCommentAvailability).toBeCalledWith(useCasePayload.commentId);
+    expect(mockThreadRepository.verifyCommentOwnership)
+      .toBeCalledWith(useCasePayload.commentId, useCasePayload.owner);
+    expect(mockThreadRepository.deleteCommentById).toBeCalledWith(useCasePayload.commentId, date);
   });
 
   it('should throw error if thread not found', async () => {
@@ -50,7 +49,7 @@ describe('DeleteThreadCommentUseCase', () => {
 
     // mock
     const mockThreadRepository = new ThreadRepository();
-    mockThreadRepository.getThreadById = jest.fn(() => Promise.resolve(false));
+    mockThreadRepository.verifyThreadAvailability = jest.fn(() => false);
 
     // Action
     const deleteThreadCommentUseCase = new DeleteThreadCommentUseCase({
@@ -61,6 +60,7 @@ describe('DeleteThreadCommentUseCase', () => {
     await expect(deleteThreadCommentUseCase.execute(useCasePayload))
       .rejects
       .toThrowError('DELETE_COMMENT_USE_CASE.THREAD_NOT_FOUND');
+    expect(mockThreadRepository.verifyThreadAvailability).toBeCalledWith(useCasePayload.threadId);
   });
 
   it('should throw error if payload not contain needed property', async () => {
@@ -91,29 +91,6 @@ describe('DeleteThreadCommentUseCase', () => {
       .toThrowError('DELETE_COMMENT_USE_CASE.NOT_MEET_DATA_TYPE_SPECIFICATION');
   });
 
-  it('should throw error when thread not found', async () => {
-    // Arrange
-    const useCasePayload = {
-      threadId: 'thread-123',
-      commentId: 'comment-123',
-      owner: 'user-123',
-    };
-
-    // mock
-    const mockThreadRepository = new ThreadRepository();
-    mockThreadRepository.getThreadById = jest.fn(() => Promise.resolve(false));
-
-    // Action
-    const deleteThreadCommentUseCase = new DeleteThreadCommentUseCase({
-      threadRepository: mockThreadRepository,
-    });
-
-    // Assert
-    await expect(deleteThreadCommentUseCase.execute(useCasePayload))
-      .rejects
-      .toThrowError('DELETE_COMMENT_USE_CASE.THREAD_NOT_FOUND');
-  });
-
   it('should throw error when comment not found', async () => {
     // Arrange
     const useCasePayload = {
@@ -124,8 +101,8 @@ describe('DeleteThreadCommentUseCase', () => {
 
     // mock
     const mockThreadRepository = new ThreadRepository();
-    mockThreadRepository.getThreadById = jest.fn(() => Promise.resolve(true));
-    mockThreadRepository.getCommentById = jest.fn(() => Promise.resolve(false));
+    mockThreadRepository.verifyThreadAvailability = jest.fn(() => true);
+    mockThreadRepository.verifyCommentAvailability = jest.fn(() => false);
 
     // Action
     const deleteThreadCommentUseCase = new DeleteThreadCommentUseCase({
@@ -136,9 +113,12 @@ describe('DeleteThreadCommentUseCase', () => {
     await expect(deleteThreadCommentUseCase.execute(useCasePayload))
       .rejects
       .toThrowError('DELETE_COMMENT_USE_CASE.COMMENT_NOT_FOUND');
+    expect(mockThreadRepository.verifyThreadAvailability).toBeCalledWith(useCasePayload.threadId);
+    expect(mockThreadRepository.verifyCommentAvailability)
+      .toBeCalledWith(useCasePayload.commentId);
   });
 
-  it('should throw error when user not the comment owner', async () => {
+  it('should throw error when not the comment owner', async () => {
     // Arrange
     const useCasePayload = {
       threadId: 'thread-123',
@@ -148,10 +128,9 @@ describe('DeleteThreadCommentUseCase', () => {
 
     // mock
     const mockThreadRepository = new ThreadRepository();
-    mockThreadRepository.getThreadById = jest.fn(() => Promise.resolve(true));
-    mockThreadRepository.getCommentById = jest.fn(() => Promise.resolve({
-      owner: 'user-321',
-    }));
+    mockThreadRepository.verifyThreadAvailability = jest.fn(() => true);
+    mockThreadRepository.verifyCommentAvailability = jest.fn(() => true);
+    mockThreadRepository.verifyCommentOwnership = jest.fn(() => false);
 
     // Action
     const deleteThreadCommentUseCase = new DeleteThreadCommentUseCase({
@@ -162,33 +141,10 @@ describe('DeleteThreadCommentUseCase', () => {
     await expect(deleteThreadCommentUseCase.execute(useCasePayload))
       .rejects
       .toThrowError('DELETE_COMMENT_USE_CASE.NOT_THE_COMMENT_OWNER');
-  });
-
-  it('should throw error when comment already deleted', async () => {
-    // Arrange
-    const useCasePayload = {
-      threadId: 'thread-123',
-      commentId: 'comment-123',
-      owner: 'user-123',
-    };
-
-    // mock
-    const mockThreadRepository = new ThreadRepository();
-    mockThreadRepository.getThreadById = jest.fn(() => Promise.resolve(true));
-    mockThreadRepository.getCommentById = jest.fn(() => Promise.resolve({
-      owner: 'user-123',
-      deletedDate: new Date().toISOString(),
-    }));
-
-    // Action
-    const deleteThreadCommentUseCase = new DeleteThreadCommentUseCase({
-      threadRepository: mockThreadRepository,
-    });
-
-    // Assert
-    await expect(deleteThreadCommentUseCase.execute(useCasePayload))
-      .rejects
-      .toThrowError('DELETE_COMMENT_USE_CASE.COMMENT_ALREADY_DELETED');
+    expect(mockThreadRepository.verifyThreadAvailability).toBeCalledWith(useCasePayload.threadId);
+    expect(mockThreadRepository.verifyCommentAvailability).toBeCalledWith(useCasePayload.commentId);
+    expect(mockThreadRepository.verifyCommentOwnership)
+      .toBeCalledWith(useCasePayload.commentId, useCasePayload.owner);
   });
 
   it('should throw error when failed to delete comment', async () => {
@@ -198,14 +154,17 @@ describe('DeleteThreadCommentUseCase', () => {
       commentId: 'comment-123',
       owner: 'user-123',
     };
+    const date = '2021-08-05T07:22:13.017Z';
 
     // mock
     const mockThreadRepository = new ThreadRepository();
-    mockThreadRepository.getThreadById = jest.fn(() => Promise.resolve(true));
-    mockThreadRepository.getCommentById = jest.fn(() => Promise.resolve({
-      owner: 'user-123',
-    }));
-    mockThreadRepository.deleteCommentById = jest.fn(() => Promise.resolve(false));
+    mockThreadRepository.verifyThreadAvailability = jest.fn(() => true);
+    mockThreadRepository.verifyCommentAvailability = jest.fn(() => true);
+    mockThreadRepository.verifyCommentOwnership = jest.fn(() => true);
+    mockThreadRepository.deleteCommentById = jest.fn(() => false);
+
+    const dateNowSpy = jest.spyOn(Date.prototype, 'toISOString');
+    dateNowSpy.mockImplementationOnce(() => date);
 
     // Action
     const deleteThreadCommentUseCase = new DeleteThreadCommentUseCase({
@@ -216,5 +175,13 @@ describe('DeleteThreadCommentUseCase', () => {
     await expect(deleteThreadCommentUseCase.execute(useCasePayload))
       .rejects
       .toThrowError('DELETE_COMMENT_USE_CASE.FAILED_TO_DELETE_COMMENT');
+    expect(mockThreadRepository.verifyThreadAvailability).toBeCalledWith(useCasePayload.threadId);
+    expect(mockThreadRepository.verifyCommentAvailability).toBeCalledWith(useCasePayload.commentId);
+    expect(mockThreadRepository.verifyCommentOwnership)
+      .toBeCalledWith(useCasePayload.commentId, useCasePayload.owner);
+    expect(mockThreadRepository.deleteCommentById).toBeCalledWith(useCasePayload.commentId, date);
+
+    // reset mock
+    dateNowSpy.mockRestore();
   });
 });
